Replace ext_tests workaround with proper integration tests
