--- ext_tests/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "js_int_ext_tests"
-version = "0.1.0"
-authors = ["Jonas Platte <jplatte+git@posteo.de>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-serde = "1.0"
-
-[dependencies.js_int]
-path = "../"
-features = ["serde"]
-
-[features]
-lax_deserialize = ["js_int/lax_deserialize"]
